[compositor] Use a tile-based system for rendering the screen
