+++ retina-compositor/Cargo.toml
@@ -11,6 +11,7 @@ retina-layout = { path = "../retina-layout" }
 retina-style = { path = "../retina-style" }
 retina-style-computation = { path = "../retina-style-computation" }
 
+crossbeam = "*"
 log = "*"
 tracing = "*"
 tracing-chrome = "*"
